+++ Localization/enUS.lua
@@ -7,7 +7,7 @@ addon.Plugins:LoadLocalization(L);
 -- [[ https://legacy.curseforge.com/wow/addons/krowi-achievement-filter/localization ]] --
 -- [[ Everything after this line is automatically generated from CurseForge and is not meant for manual edit - SOURCETOKEN - AUTOGENTOKEN ]] --
 
--- [[ Exported at 2023-10-30 16-54-38 ]] --
+-- [[ Exported at 2023-11-04 19-45-07 ]] --
 L["%c"] = true
 L["%d/%m/%Y %I:%M %p"] = true
 L["%m/%d/%Y %I:%M %p"] = true
@@ -879,6 +879,7 @@ L["Summary Desc"] = [=[{achievementsHeader}:
 L["Sundapple Copse Circuit"] = true
 L["Superbloom"] = true
 L["Superbloom: Active"] = true
+L["Superbloom: Next"] = true
 L["Tab"] = true
 L["tab"] = true
 L["Tabs"] = true